Move DataLoadState and LoadState intefaces to new file
