+++ tensorboard/webapp/plugins/plugins_container.ts
@@ -16,7 +16,8 @@ import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {Store, select, createSelector} from '@ngrx/store';
 
 import {getPlugins, getActivePlugin, getPluginsListLoaded} from '../core/store';
-import {LoadState, PluginMetadata} from '../types/api';
+import {PluginMetadata} from '../types/api';
+import {LoadState} from '../types/data';
 import {State} from '../core/store/core_types';
 
 /** @typehack */ import * as _typeHackRxjs from 'rxjs';
