+++ tensorboard/webapp/types/api.ts
@@ -13,19 +13,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-export enum DataLoadState {
-  NOT_LOADED,
-  LOADED,
-  LOADING,
-  FAILED,
-}
-
-export interface LoadState {
-  state: DataLoadState;
-  // Time since epoch.
-  lastLoadedTimeInMs: number | null;
-}
-
 export type PluginId = string;
 
 export enum LoadingMechanismType {
